+++ src/main.js
@@ -1,6 +1,7 @@
 import Vue from "vue";
 import axios from "axios";
 import VueAxios from "vue-axios";
+import { Auth0Plugin } from "./auth";
 
 import App from "./App.vue";
 import "./registerServiceWorker";
@@ -9,10 +10,7 @@ import router from "./router";
 // Import the Auth0 configuration
 import { domain, clientId } from "../auth_config.json";
 
-// Import the plugin here
-import { Auth0Plugin } from "./auth";
-
-Vue.use(VueAxios, axios, Auth0Plugin, {
+Vue.use(Auth0Plugin, {
   domain,
   clientId,
   onRedirectCallback: appState => {
@@ -23,6 +21,8 @@ Vue.use(VueAxios, axios, Auth0Plugin, {
     );
   }
 });
+
+Vue.use(VueAxios, axios);
 
 Vue.config.productionTip = false;
